refactor: study detail 페이지 코드  경로  수정
