+++ src/pages/study-create/StudyCreatePage.jsx
@@ -3,8 +3,8 @@ import { useNavigate } from 'react-router-dom';
 import styled from 'styled-components';
 import backGroundUrl from '../../assets/images/mypage/mypageBackground.png';
 import StudyInfo from './ui/StudyInfo';
-import StudyPeriod from './ui/StudyInfo';
-import StudyDetail from './ui/StudyInfo';
+import StudyPeriod from './ui/StudyPeriod';
+import StudyDetail from './ui/StudyDetail';
 import { studyCreateAPI } from './api/studyCreateAPI';
 import { useSelector } from 'react-redux';
 import { ContentWrapper } from '../../components/common/MediaWrapper';
